Inference reorg
