+++ examples/log_parsing/inference.py
@@ -30,8 +30,8 @@ from scipy.special import softmax
 from morpheus.config import Config
 from morpheus.stages.inference.inference_stage import InferenceStage
 from morpheus.stages.inference.inference_stage import InferenceWorker
-from morpheus.stages.inference.inference_triton import InputWrapper
-from morpheus.stages.inference.inference_triton import TritonInferenceWorker
+from morpheus.stages.inference.triton_inference_stage import InputWrapper
+from morpheus.stages.inference.triton_inference_stage import _TritonInferenceWorker
 from morpheus.messages import InferenceMemory
 from morpheus.messages import MultiInferenceMessage
 from morpheus.pipeline.pipeline import StreamPair
@@ -40,7 +40,7 @@ from morpheus.utils.producer_consumer_queue import ProducerConsumerQueue
 logger = logging.getLogger(__name__)
 
 
-class TritonInferenceLogParsing(TritonInferenceWorker):
+class TritonInferenceLogParsing(_TritonInferenceWorker):
     """
     This class extends TritonInference to deal with scenario-specific NLP models inference requests like building
     response.
