+++ tests/test_phishing.py
@@ -24,7 +24,7 @@ from morpheus.config import PipelineModes
 from morpheus.pipeline import LinearPipeline
 from morpheus.stages.general.general_stages import AddClassificationsStage
 from morpheus.stages.general.general_stages import MonitorStage
-from morpheus.stages.inference.inference_triton import TritonInferenceStage
+from morpheus.stages.inference.triton_inference_stage import TritonInferenceStage
 from morpheus.stages.input.file_source_stage import FileSourceStage
 from morpheus.stages.postprocess.serialize_stage import SerializeStage
 from morpheus.stages.output.write_to_file_stage import WriteToFileStage
